# perf(core): use BTreeMap for GothamState